+++ petab_select/constants.py
@@ -1,7 +1,8 @@
 """Constants for the PEtab Select package."""
+import sys
 from enum import Enum
 from pathlib import Path
-from typing import Dict, List, Union
+from typing import Dict, List, Literal, Union
 
 # Zero-indexed column/row indices
 MODEL_ID_COLUMN = 0
@@ -48,12 +49,6 @@ MODEL_SPACE_FILE_NON_PARAMETER_COLUMNS = [MODEL_SUBSPACE_ID, PETAB_YAML]
 # COMPARED_MODEL_ID = 'compared_'+MODEL_ID
 YAML_FILENAME = 'yaml'
 
-# FORWARD = 'forward'
-# BACKWARD = 'backward'
-# BIDIRECTIONAL = 'bidirectional'
-# LATERAL = 'lateral'
-
-
 # DISTANCES = {
 #    FORWARD: {
 #        'l1': 1,
@@ -70,20 +65,6 @@ YAML_FILENAME = 'yaml'
 # }
 
 CRITERIA = 'criteria'
-# FIXME remove, change all uses to Enum below
-# AIC = 'AIC'
-# AICC = 'AICc'
-# BIC = 'BIC'
-# AKAIKE_INFORMATION_CRITERION = AIC
-# CORRECTED_AKAIKE_INFORMATION_CRITERION = AICC
-# BAYESIAN_INFORMATION_CRITERION = BIC
-# LH = 'LH'
-# LLH = 'LLH'
-# NLLH = 'NLLH'
-# LIKELIHOOD = LH
-# LOG_LIKELIHOOD = LLH
-# NEGATIVE_LOG_LIKELIHOOD = NLLH
-
 
 PARAMETERS = 'parameters'
 # PARAMETER_ESTIMATE = 'parameter_estimate'
@@ -105,8 +86,7 @@ MODELS = 'models'
 
 # Parameters can be fixed to a value, or estimated if indicated with the string
 # `ESTIMATE`.
-# TODO change to `Literal[ESTIMATE]` (Python >= 3.8)
-TYPE_PARAMETER = Union[float, int, ESTIMATE]
+TYPE_PARAMETER = Union[float, int, Literal[ESTIMATE]]
 TYPE_PARAMETER_OPTIONS = List[TYPE_PARAMETER]
 # Parameter ID -> parameter value mapping.
 TYPE_PARAMETER_DICT = Dict[str, TYPE_PARAMETER]
@@ -119,49 +99,62 @@ TYPE_CRITERION = float
 class Method(str, Enum):
     """String literals for model selection methods."""
 
+    #: The backward stepwise method.
     BACKWARD = 'backward'
-    BIDIRECTIONAL = 'bidirectional'
+    #: The brute-force method.
     BRUTE_FORCE = 'brute_force'
+    #: The FAMoS method.
     FAMOS = 'famos'
+    #: The forward stepwise method.
     FORWARD = 'forward'
-    FORWARD_AND_BACKWARD = 'forward_and_backward'
+    #: The lateral, or swap, method.
     LATERAL = 'lateral'
+    #: The jump-to-most-distant-model method.
     MOST_DISTANT = 'most_distant'
 
 
 class Criterion(str, Enum):
     """String literals for model selection criteria."""
 
+    #: The Akaike information criterion.
     AIC = 'AIC'
+    #: The corrected Akaike information criterion.
     AICC = 'AICc'
+    #: The Bayesian information criterion.
     BIC = 'BIC'
+    #: The likelihood.
     LH = 'LH'
+    #: The log-likelihood.
     LLH = 'LLH'
+    #: The negative log-likelihood.
     NLLH = 'NLLH'
 
 
-# Methods that move through model space by taking steps away from some model.
+#: Methods that move through model space by taking steps away from some model.
 STEPWISE_METHODS = [
     Method.BACKWARD,
-    Method.BIDIRECTIONAL,
     Method.FORWARD,
-    Method.FORWARD_AND_BACKWARD,
     Method.LATERAL,
 ]
-# Methods that require an initial model.
+#: Methods that require an initial model.
 INITIAL_MODEL_METHODS = [
     Method.BACKWARD,
-    Method.BIDIRECTIONAL,
     Method.FORWARD,
-    Method.FORWARD_AND_BACKWARD,
     Method.LATERAL,
 ]
 
-# Virtual initial models can be used to initialize some initial model methods.
+#: Virtual initial models can be used to initialize some initial model methods.
 VIRTUAL_INITIAL_MODEL = 'virtual_initial_model'
+#: Methods that are compatible with a virtual initial model.
 VIRTUAL_INITIAL_MODEL_METHODS = [
     Method.BACKWARD,
-    Method.BIDIRECTIONAL,
     Method.FORWARD,
-    Method.FORWARD_AND_BACKWARD,
+]
+
+
+__all__ = [
+    x
+    for x in dir(sys.modules[__name__])
+    if not x.startswith('_')
+    and x not in ('sys', "Enum", "Path", "Dict", "List", "Literal", "Union")
 ]