+++ petab_select/model.py
@@ -1,11 +1,9 @@
 """The `Model` class."""
-import abc
 import warnings
 from os.path import relpath
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Tuple, Union
 
-import numpy as np
 import petab
 import yaml
 from more_itertools import one
@@ -38,6 +36,13 @@ from .misc import (
 )
 from .petab import PetabMixin
 
+__all__ = [
+    'Model',
+    'default_compare',
+    'models_from_yaml_list',
+    'models_to_yaml_list',
+]
+
 
 class Model(PetabMixin):
     """A (possibly uncalibrated) model.
@@ -48,16 +53,16 @@ class Model(PetabMixin):
 
     Attributes:
         converters_load:
-            Functions to convert attributes from YAML to `Model`.
+            Functions to convert attributes from YAML to :class:`Model`.
         converters_save:
-            Functions to convert attributes from `Model` to YAML.
+            Functions to convert attributes from :class:`Model` to YAML.
         criteria:
             The criteria values of the calibrated model (e.g. AIC).
         hash_attributes:
             This attribute is currently not used.
             Attributes that will be used to calculate the hash of the
-            `Model` instance. NB: this hash is used during pairwise comparison
-            to determine whether any two `Model` instances are unique. The
+            :class:`Model` instance. NB: this hash is used during pairwise comparison
+            to determine whether any two :class:`Model` instances are unique. The
             model instances are compared by their parameter estimation
             problems, as opposed to parameter estimation results, which may
             differ due to e.g. floating-point arithmetic.
@@ -71,9 +76,10 @@ class Model(PetabMixin):
         estimated_parameters:
             Parameter estimates from a model calibration tool, for parameters
             that are specified as estimated in the PEtab problem or PEtab
-            Select model YAML.
+            Select model YAML. These are untransformed values (i.e., not on
+            log scale).
         saved_attributes:
-            Attributes that will be saved to disk by the `Model.to_yaml`
+            Attributes that will be saved to disk by the :meth:`Model.to_yaml`
             method.
     """
 
@@ -99,7 +105,7 @@ class Model(PetabMixin):
         ESTIMATED_PARAMETERS: lambda x: x,
         CRITERIA: lambda x: {
             # `criterion_id_value` is the ID of the criterion in the enum `Criterion`.
-            Criterion(criterion_id_value): criterion_value
+            Criterion(criterion_id_value): float(criterion_value)
             for criterion_id_value, criterion_value in x.items()
         },
     }
@@ -108,7 +114,7 @@ class Model(PetabMixin):
         MODEL_SUBSPACE_ID: lambda x: str(x),
         MODEL_SUBSPACE_INDICES: lambda x: [int(xi) for xi in x],
         MODEL_HASH: lambda x: str(x),
-        PREDECESSOR_MODEL_HASH: lambda x: str(x),
+        PREDECESSOR_MODEL_HASH: lambda x: str(x) if x is not None else x,
         PETAB_YAML: lambda x: str(x),
         PARAMETERS: lambda x: {str(k): v for k, v in x.items()},
         # FIXME handle with a `set_estimated_parameters` method instead?
@@ -120,7 +126,7 @@ class Model(PetabMixin):
             str(id): float(value) for id, value in x.items()
         },
         CRITERIA: lambda x: {
-            criterion_id.value: criterion_value
+            criterion_id.value: float(criterion_value)
             for criterion_id, criterion_value in x.items()
         },
     }
@@ -191,7 +197,7 @@ class Model(PetabMixin):
 
         Args:
             criterion:
-                The criterion (e.g. `petab_select.constants.Criterion.AIC`).
+                The criterion (e.g. ``petab_select.constants.Criterion.AIC``).
             value:
                 The criterion value for the (presumably calibrated) model.
         """
@@ -223,45 +229,87 @@ class Model(PetabMixin):
         self,
         criterion: Criterion,
         compute: bool = True,
+        raise_on_failure: bool = True,
     ) -> Union[TYPE_CRITERION, None]:
         """Get a criterion value for the model.
 
         Args:
             criterion:
-                The ID of the criterion (e.g. `petab_select.constants.Criterion.AIC`).
+                The ID of the criterion (e.g. ``petab_select.constants.Criterion.AIC``).
             compute:
                 Whether to try to compute the criterion value based on other model
-                attributes. For example, if the `'AIC'` criterion is requested, this
+                attributes. For example, if the ``'AIC'`` criterion is requested, this
                 can be computed from a predetermined model likelihood and its
                 number of estimated parameters.
+            raise_on_failure:
+                Whether to raise a `ValueError` if the criterion could not be
+                computed. If `False`, `None` is returned.
 
         Returns:
             The criterion value, or `None` if it is not available.
             TODO check for previous use of this method before `.get` was used
         """
         if criterion not in self.criteria and compute:
-            self.compute_criterion(criterion=criterion)
+            self.compute_criterion(
+                criterion=criterion,
+                raise_on_failure=raise_on_failure,
+            )
             # value = self.criterion_computer(criterion=id)
             # self.set_criterion(id=id, value=value)
 
         return self.criteria.get(criterion, None)
 
-    def compute_criterion(self, criterion: Criterion) -> TYPE_CRITERION:
+    def compute_criterion(
+        self,
+        criterion: Criterion,
+        raise_on_failure: bool = True,
+    ) -> TYPE_CRITERION:
         """Compute a criterion value for the model.
 
         The value will also be stored, which will overwrite any previously stored value
         for the criterion.
 
         Args:
-            id:
-                The ID of the criterion (e.g. `petab_select.constants.Criterion.AIC`).
+            criterion:
+                The ID of the criterion
+                (e.g. :obj:`petab_select.constants.Criterion.AIC`).
+            raise_on_failure:
+                Whether to raise a `ValueError` if the criterion could not be
+                computed. If `False`, `None` is returned.
 
         Returns:
             The criterion value.
         """
-        criterion_value = self.criterion_computer(criterion)
-        self.set_criterion(criterion, criterion_value)
-        return criterion_value
+        try:
+            criterion_value = self.criterion_computer(criterion)
+            self.set_criterion(criterion, criterion_value)
+            result = criterion_value
+        except ValueError:
+            if raise_on_failure:
+                raise
+            result = None
+        return result
+
+    def set_estimated_parameters(
+        self,
+        estimated_parameters: Dict[str, float],
+        scaled: bool = False,
+    ) -> None:
+        """Set the estimated parameters.
+
+        Args:
+            estimated_parameters:
+                The estimated parameters.
+            scaled:
+                Whether the ``estimated_parameters`` values are on the scale
+                defined in the PEtab problem (``True``), or untransformed
+                (``False``).
+        """
+        if scaled:
+            estimated_parameters = self.petab_problem.unscale_parameters(
+                estimated_parameters
+            )
+        self.estimated_parameters = estimated_parameters
 
     @staticmethod
     def from_dict(
@@ -276,15 +324,15 @@ class Model(PetabMixin):
                 A dictionary of attributes. The keys are attribute
                 names, the values are the corresponding attribute values for
                 the model. Required attributes are the required arguments of
-                the `Model.__init__` method.
+                the :meth:`Model.__init__` method.
             base_path:
                 The path that any relative paths in the model are relative to
                 (e.g. the path to the PEtab problem YAML file
-                `Model.petab_yaml` may be relative).
+                :meth:`Model.petab_yaml` may be relative).
             petab_problem:
                 Optionally provide the PEtab problem, to avoid loading it multiple
                 times.
-                NB: This may causes issues if multiple models write to the same PEtab
+                NB: This may cause issues if multiple models write to the same PEtab
                 problem in memory.
 
         Returns:
@@ -341,20 +389,20 @@ class Model(PetabMixin):
         resolve_paths: bool = True,
         paths_relative_to: Union[str, Path] = None,
     ) -> Dict[str, Any]:
-        """Generate a dictionary from the attributes of a `Model` instance.
+        """Generate a dictionary from the attributes of a :class:`Model` instance.
 
         Args:
             resolve_paths:
                 Whether to resolve relative paths into absolute paths.
             paths_relative_to:
-                If not `None`, paths will be converted to be relative to this path.
-                Takes priority over `resolve_paths`.
+                If not ``None``, paths will be converted to be relative to this path.
+                Takes priority over ``resolve_paths``.
 
         Returns:
             A dictionary of attributes. The keys are attribute
             names, the values are the corresponding attribute values for
             the model. Required attributes are the required arguments of
-            the `Model.__init__` method.
+            the :meth:`Model.__init__` method.
         """
         model_dict = {}
         for attribute in self.saved_attributes:
@@ -376,14 +424,14 @@ class Model(PetabMixin):
         return model_dict
 
     def to_yaml(self, petab_yaml: TYPE_PATH, *args, **kwargs) -> None:
-        """Generate a PEtab Select model YAML file from a `Model` instance.
+        """Generate a PEtab Select model YAML file from a :class:`Model` instance.
 
         Parameters:
             petab_yaml:
                 The location where the PEtab Select model YAML file will be
                 saved.
             args, kwargs:
-                Additional arguments are passed to `self.to_dict`.
+                Additional arguments are passed to ``self.to_dict``.
         """
         # FIXME change `getattr(self, PETAB_YAML)` to be relative to
         # destination?
@@ -395,27 +443,50 @@ class Model(PetabMixin):
     def to_petab(
         self,
         output_path: TYPE_PATH = None,
-    ) -> Tuple[petab.Problem, TYPE_PATH]:
+        set_estimated_parameters: Optional[bool] = None,
+    ) -> Dict[str, Union[petab.Problem, TYPE_PATH]]:
         """Generate a PEtab problem.
 
         Args:
             output_path:
                 The directory where PEtab files will be written to disk. If not
                 specified, the PEtab files will not be written to disk.
+            set_estimated_parameters:
+                Whether to set the nominal value of estimated parameters to their
+                estimates. If parameter estimates are available, this
+                will default to `True`.
 
         Returns:
             A 2-tuple. The first value is a PEtab problem that can be used
             with a PEtab-compatible tool for calibration of this model. If
-            `output_path` is not `None`, the second value is the path to a
+            ``output_path`` is not ``None``, the second value is the path to a
             PEtab YAML file that can be used to load the PEtab problem (the
-            first value) into any PEtab-compatible tool. If
+            first value) into any PEtab-compatible tool.
         """
         # TODO could use `copy.deepcopy(self.petab_problem)` from PetabMixin?
         petab_problem = petab.Problem.from_yaml(str(self.petab_yaml))
+
+        if set_estimated_parameters is None and self.estimated_parameters:
+            set_estimated_parameters = True
+
         for parameter_id, parameter_value in self.parameters.items():
             # If the parameter is to be estimated.
             if parameter_value == ESTIMATE:
                 petab_problem.parameter_df.loc[parameter_id, ESTIMATE] = 1
+
+                if set_estimated_parameters:
+                    if parameter_id not in self.estimated_parameters:
+                        raise ValueError(
+                            "Not all estimated parameters are available "
+                            "in `model.estimated_parameters`. Hence, the "
+                            "estimated parameter vector cannot be set as "
+                            "the nominal value in the PEtab problem. "
+                            "Try calling this method with "
+                            "`set_estimated_parameters=False`."
+                        )
+                    petab_problem.parameter_df.loc[
+                        parameter_id, NOMINAL_VALUE
+                    ] = self.estimated_parameters[parameter_id]
             # Else the parameter is to be fixed.
             else:
                 petab_problem.parameter_df.loc[parameter_id, ESTIMATE] = 0
@@ -444,7 +515,7 @@ class Model(PetabMixin):
         is calibrated twice and the two calibrated models differ in their parameter
         estimates, then they will still have the same hash.
 
-        This is not implemented as `__hash__` because Python automatically truncates
+        This is not implemented as ``__hash__`` because Python automatically truncates
         values in a system-dependent manner, which reduces interoperability
         ( https://docs.python.org/3/reference/datamodel.html#object.__hash__ ).
 
@@ -481,8 +552,9 @@ class Model(PetabMixin):
         return f'{header}\n{data}'
 
     def get_mle(self) -> Dict[str, float]:
-        """Get the maximum likelihood estimate of the model.
-
+        """Get the maximum likelihood estimate of the model."""
+        """
+        FIXME(dilpath)
         # Check if original PEtab problem or PEtab Select model has estimated
         # parameters. e.g. can use some of `self.to_petab` to get the parameter
         # df and see if any are estimated.
@@ -554,10 +626,10 @@ class Model(PetabMixin):
     ) -> List[TYPE_PARAMETER]:
         """Get parameter values.
 
-        Includes `ESTIMATE` for parameters that should be estimated.
+        Includes ``ESTIMATE`` for parameters that should be estimated.
 
-        The ordering is by `parameter_ids` if supplied, else
-        `self.petab_parameters`.
+        The ordering is by ``parameter_ids`` if supplied, else
+        ``self.petab_parameters``.
 
         Args:
             parameter_ids:
@@ -587,8 +659,8 @@ def default_compare(
 ) -> bool:
     """Compare two calibrated models by their criterion values.
 
-    It is assumed that the model `model0` provides a value for the criterion
-    `criterion`, or is the `VIRTUAL_INITIAL_MODEL`.
+    It is assumed that the model ``model0`` provides a value for the criterion
+    ``criterion``, or is the ``VIRTUAL_INITIAL_MODEL``.
 
     Args:
         model0:
@@ -602,8 +674,8 @@ def default_compare(
             model. Should be non-negative.
 
     Returns:
-        `True` if `model1` has a better criterion value than `model0`, else
-        `False`.
+        ``True` if ``model1`` has a better criterion value than ``model0``, else
+        ``False``.
     """
     if not model1.has_criterion(criterion):
         warnings.warn(
@@ -650,11 +722,11 @@ def models_from_yaml_list(
         model_list_yaml:
             The path to the PEtab Select list of model YAML file.
         petab_problem:
-            See `Model.from_dict`.
+            See :meth:`Model.from_dict`.
         allow_single_model:
             Given a YAML file that contains a single model directly (not in
-            a 1-element list), if `True` then the single model will be read in,
-            else an error will be raised.
+            a 1-element list), if ``True`` then the single model will be read in,
+            else a ``ValueError`` will be raised.
 
     Returns:
         A list of model instances, initialized with the provided