+++ petab_select/model_space.py
@@ -1,40 +1,33 @@
 """The `ModelSpace` class and related methods."""
-import abc
 import itertools
 import logging
+import warnings
 from pathlib import Path
 from tempfile import NamedTemporaryFile
-from typing import (
-    Any,
-    Callable,
-    Iterable,
-    List,
-    Optional,
-    TextIO,
-    Union,
-    get_args,
-)
+from typing import Any, Iterable, List, Optional, TextIO, Union, get_args
 
 import numpy as np
 import pandas as pd
-from more_itertools import nth
 
 from .candidate_space import CandidateSpace
 from .constants import (
-    ESTIMATE,
     HEADER_ROW,
-    MODEL_ID,
     MODEL_ID_COLUMN,
-    MODEL_SPACE_FILE_NON_PARAMETER_COLUMNS,
     MODEL_SUBSPACE_ID,
     PARAMETER_DEFINITIONS_START,
     PARAMETER_VALUE_DELIMITER,
-    PETAB_YAML,
     PETAB_YAML_COLUMN,
     TYPE_PATH,
 )
 from .model import Model
 from .model_subspace import ModelSubspace
+
+__all__ = [
+    "ModelSpace",
+    "get_model_space_df",
+    "read_model_space_file",
+    "write_model_space_df",
+]
 
 
 def read_model_space_file(filename: str) -> TextIO:
@@ -49,7 +42,9 @@ def read_model_space_file(filename: str) -> TextIO:
 
     Returns:
         A temporary file object, which is the unpacked file.
-
+    """
+    """
+    FIXME(dilpath)
     Todo:
         * Consider alternatives to `_{n}` suffix for model `modelId`
         * How should the selected model be reported to the user? Remove the
@@ -113,13 +108,13 @@ def line2row(
         delimiter:
             The string that separates columns in the file.
         unpacked:
-            Whether the line format is in the unpacked format. If False,
-            parameter values are not converted to `float`.
+            Whether the line format is in the unpacked format. If ``False``,
+            parameter values are not converted to ``float``.
         convert_parameters_to_float:
-            Whether parameters should be converted to `float`.
+            Whether parameters should be converted to ``float``.
 
     Returns:
-        A list of column values. Parameter values are converted to `float`.
+        A list of column values. Parameter values are converted to ``float``.
     """
     columns = line.strip().split(delimiter)
     metadata = columns[:PARAMETER_DEFINITIONS_START]
@@ -265,7 +260,7 @@ class ModelSpace:
 
     def __len__(self):
         """Get the number of models in this space."""
-        subspace_coumts = [len(s) for s in self.model_subspaces]
+        subspace_counts = [len(s) for s in self.model_subspaces]
         total_count = sum(subspace_counts)
         return total_count